base-hw: don't use 'long' unnecessarily
Fix #581

+++ base-hw/include/base/ipc_pager.h
@@ -151,7 +151,7 @@ namespace Genode
 			/**
 			 * Return badge for faulting thread
 			 */
-			unsigned long badge() const { return _pagefault.thread_id; }
+			unsigned badge() const { return _pagefault.thread_id; }
 
 			/**
 			 * Return true if last fault was a write fault
