+++ base-hw/include/signal_session/signal_session.h
@@ -76,7 +76,7 @@ namespace Genode
 		 */
 		virtual Signal_context_capability
 		alloc_context(Signal_receiver_capability const r,
-		              unsigned long const imprint) = 0;
+		              unsigned const imprint) = 0;
 
 		/*********************
 		 ** RPC declaration **
@@ -86,7 +86,7 @@ namespace Genode
 		                 alloc_receiver, GENODE_TYPE_LIST(Out_of_metadata));
 		GENODE_RPC_THROW(Rpc_alloc_context, Signal_context_capability,
 		                 alloc_context, GENODE_TYPE_LIST(Out_of_metadata),
-		                 Signal_receiver_capability, unsigned long);
+		                 Signal_receiver_capability, unsigned);
 
 		GENODE_RPC_INTERFACE(Rpc_alloc_receiver, Rpc_alloc_context);
 	};
